export constant channel & moved phone validation method to validate_data.py
