+++ app/core/services/SMS/TwilioVerifyService.py
@@ -43,7 +43,6 @@ class TwilioVerifyService(SMSService):
             verify_service_sid or settings.TWILIO_VERIFY_SERVICE_SID
         )
         self.verify = self.client.verify.services(self.verify_service_sid)
-        self.channel = "sms"
 
     def send_verification_code(self, phone):
         """
@@ -57,11 +56,11 @@ class TwilioVerifyService(SMSService):
         """
         try:
             verification = self.verify.verifications.create(
-                to=phone, channel=self.channel
+                to=phone, channel=settings.CHANNEL
             )
             return verification.sid
-        except TwilioRestException as e:
-            raise SMSException(f"Error sending verification code: {e.msg}")
+        except TwilioRestException:
+            raise SMSException(f"Error sending verification code")
 
     def check_verification_code(self, phone, code):
         """
@@ -78,4 +77,4 @@ class TwilioVerifyService(SMSService):
             result = self.verify.verification_checks.create(to=phone, code=code)
             return result.status == "approved"
         except TwilioRestException as e:
-            raise SMSException(f"Error receiving verification code: {e.msg}")
+            raise SMSException(f"Error receiving verification code")
