+++ project_core/settings/sms_sending.py
@@ -8,3 +8,4 @@ env = environ.Env()
 TWILIO_ACCOUNT_SID = env("TWILIO_ACCOUNT_SID")
 TWILIO_AUTH_TOKEN = env("TWILIO_AUTH_TOKEN")
 TWILIO_VERIFY_SERVICE_SID = env("TWILIO_VERIFY_SERVICE_SID")
+CHANNEL = "sms"
